fix: remove user groups file, fix role dropdown styling
