+++ src/Pages/GlobalConfigurations/Authorization/shared/components/K8sObjectPermissions/K8sListItemCard.tsx
@@ -40,11 +40,10 @@ import { k8sRoleSelectionStyle, multiSelectAllState } from './K8sPermissions.uti
 import { resourceKindOptionLabel } from './K8sPermission.component'
 import { useAuthorizationContext } from '../../../AuthorizationProvider'
 import { parseData } from '../../../utils'
-import { authorizationSelectStyles } from '../userGroups/UserGroup'
+import { authorizationSelectStyles, ALL_NAMESPACE, EntityTypes } from '../../../constants'
 import { K8sPermissionActionType, K8S_PERMISSION_INFO_MESSAGE } from './constants'
 import { SELECT_ALL_VALUE } from '../../../../../../config'
 import { K8sItemCardLoadingState, K8sListItemCardType } from './types'
-import { ALL_NAMESPACE, EntityTypes } from '../../../constants'
 import { usePermissionConfiguration } from '../PermissionConfigurationForm'
 import { K8sPermissionFilter } from '../../../types'
 import { getIsStatusDropdownDisabled } from '../../../libUtils'
