+++ src/Pages/GlobalConfigurations/Authorization/shared/components/K8sObjectPermissions/K8sPermissions.utils.ts
@@ -1,9 +1,8 @@
 import { ReactSelectInputAction } from '@devtron-labs/devtron-fe-common-lib'
 import { SELECT_ALL_VALUE } from '../../../../../../config'
-import { ActionTypes, ACTION_LABEL, EntityTypes } from '../../../constants'
+import { ActionTypes, ACTION_LABEL, EntityTypes, authorizationSelectStyles } from '../../../constants'
 import { K8sPermissionFilter } from '../../../types'
 import { getDefaultStatusAndTimeout } from '../../../libUtils'
-import { authorizationSelectStyles } from '../userGroups/UserGroup'
 
 export const apiGroupAll = (permission, isLabel = false) => {
     if (permission === '') {
